class22: start test_derivative
It's using xtensor, so I'm using the same cmake stuff to make it available.

For now, test_derivative doesn't do much yet, it makes coordinates from 2pi
and creates the function I wnat to take a derivative of.

It writes a csv file, which can be plotted in gnuplot without too much
trouble, see the included script